Rework the CPU address space.
What was implemented as the NES address space was actually the CPU
address space.

This refactoring will make it more natural to have the PPU address space
as well.

Unfortunately, due to Rust's ownership rules, it's hard to move these
address spaces in the struct themselve (i.e.: have an AddressSpace field
on Cpu) because they refer back to Nes which makes them
self-referencing.

It's certainly possible, not sure how.

The solution is to build the address spaces when the program runs; e.g.:
wiring everything together only when run is called and tearing
everything down after.