Even more ppu nmi tweaks.
This makes test 07 fail now, but both 05 and 06 are one clock cycle closer to passing, so this sounds like a win.

At this point, I suspect the issue might be in the cpu / ppu sync which is off due to cpu timing being not cycle perfect.
Checking cpu timing will require a partial implementation of the APU, so I might just ignore this for now.