Stepper refactored to consume Nes.
This removes a TODO, but uses lots of unsafety to do so.

The upside is that Nes is consumed by the Stepper which makes sense
since you shouldn't be stepping the same Nes from 2 places.

Perhaps a future refactoring is to make Stepper be the Nes itself and
push Nes down a level of abstraction to something along the lines of
"hardware" or similar.